requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
crispy-bootstrap5==2025.4
cryptography==45.0.3
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
Django==4.2.21
django-admin-thumbnails==0.2.9
django-allauth==65.9.0
django-crispy-forms==2.4
django-redis==6.0.0
dotenv==0.9.9
elevenlabs==2.6.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.16.1
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-tavily==0.2.5
langchain-text-splitters==0.3.8
langchain-together==0.3.0
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langgraph-supervisor==0.0.27
langsmith==0.4.4
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
openai==1.93.0
orjson==3.10.15
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==27.0.0
razorpay==1.4.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
together==1.3.14
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
typing_extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
